+++ packages/core/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## next (YYYY-MM-DD)
 
-* Feature: auto alias for objects and inputs
+* Feature: auto assign names to objects and inputs
 
 ## 0.5.0 (2024-12-03)